+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support `float` and `double` flags.
+
 ## [1.0.0] - 2020-08-18
 ### Changed
 - Remove `typedef` for types `CFlag` and `CFlagStatus`, instead user code
@@ -21,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
+[Unreleased]: https://github.com/aperezdc/cflag/compare/1.0.0...HEAD
 [1.0.0]: https://github.com/aperezdc/cflag/compare/0.1.1...1.0.0
 [0.1.1]: https://github.com/aperezdc/cflag/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/aperezdc/cflag/releases/tag/0.1.0